Move the user fetch MW to Middlewares folder
